fix various compile errors

+++ GURPS Star System Generator/Star.h
@@ -18,11 +18,11 @@ private:
 	double stellar_mass{ 0 };			// expressed in solar masses
 	std::string spectral_type{ "" };	// spectral type of star in main sequence e.g. G2
 	int temp{ 0 };						// surface temperature in kelvins
-	float l_max{ 0 };					// maximum luminosity in main sequence ** -1 if not applicable **
-	float l_min{ 0 };					// minimum luminosity in main sequence
-	float m_span{ 0 };					// duration of main sequence in billions of years ** -1 if not applicable **
-	float s_span{ 0 };					// duration of sub-giant phase in billions of years ** -1 if not applicable **
-	float g_span{ 0 };					// duration of giant phase in billions of years ** -1 if not applicable **
+	double l_max{ 0 };					// maximum luminosity in main sequence ** -1 if not applicable **
+	double l_min{ 0 };					// minimum luminosity in main sequence
+	double m_span{ 0 };					// duration of main sequence in billions of years ** -1 if not applicable **
+	double s_span{ 0 };					// duration of sub-giant phase in billions of years ** -1 if not applicable **
+	double g_span{ 0 };					// duration of giant phase in billions of years ** -1 if not applicable **
 
 public:
 	//**************
@@ -38,11 +38,11 @@ public:
 	bool get_Has_Companion();		// does this star have a companion?
 	int get_Num_Of_Companions();	// output number of companions
 	bool get_Is_Companion();		// is this star a companion?
-	float get_L_Max();				// output l_max
-	float get_L_Min();				// output l-min
-	float get_M_Span();				// output phase lifespans
-	float get_S_Span();				
-	float get_G_Span();
+	double get_L_Max();				// output l_max
+	double get_L_Min();				// output l-min
+	double get_M_Span();				// output phase lifespans
+	double get_S_Span();				
+	double get_G_Span();
 	double get_Mass();				// output value of stellar_mass
 	Star* get_Primary();			// output pointer to primary star or nullptr
 	std::string get_Spectral_Type();// output spectral type
